Fixed some of the merge and parsing code.
Transitioned to a pipeline appraoch, to restrict I/O to one function.
I missed some matchers, and didn't realize that the sequence pattern
matching was a pop, rather than just matching out the head or tail.
In retrospect, it's obvious (since you can do it more than once).
Also refactored to build a solver and board gegenerator executable
in the same project.