Implemented Scorched Earth Kernel, Build/Move func
Implemented Build and Move functions that transform IBoards. Need some
solid unit testing later.

Implemented a "Scorched Earth" pure kernel, which just picks the first
player that has moves, moves it to the first legal location, then builds
in the location we were before. Intended to validate game logic rules.

Moved height from Rep to Logic, since it's easiest to implement in terms
of tokens. Ultimately, a bunch of stuff might move in the other
direction, but that requires re-arranging some unit tests, and it's not
worth the time right now.

NOTE: Content from this commit comes from the selected answer to this
      StackOverflow question. Content is marked with a similar citation.
      https://stackoverflow.com/questions/5852722/replace-individual-list-elements-in-haskell
      Answered by username Philip JF, edited by user adius