Implemented pushPlayer
Implemented pushPlayer, and tweaked the unit tests to match the fact
that, when updating a player, the updated token comes first. If I ever
have to touch this code again, I ought to define a custom equality
function for players that ignores token order, but this got it working
ASAP. Also added a few short functions, mostly simple getters, some
error checking, some not.