Implemented isWon for cards
Took a shortcut, but it's one that I think is acceptable. The W/L
function for Minotaur checks to see if the other player is Pan, and then
checks the Pan win condition. A better technique would be to make sure
we didn't lose as the result of our move (i.e: that the other player
didn't Win), but it doesn't seem like that'd be useful for any other
combination.